Fourteenth Assignment in Software Development
This is my fourteenth assignment in my software development class. This is a game where the rocket has to dodge the asteroids, if it is unsuccessful, then the user has lost the game.
In order to view, play, etc. you have to have the software Greenfoot. This is a CPU demanding software as it uses 30% of the storage so, install it with caution and be aware of the possible repercussions/consequences/outcomes.